Split Telegram channel into focused modules
Move Telegram DTOs, state paths, status mapping, polling, sending,
webhook handling, attachments, downloads, inbound dispatch, guest
entrypoints, and tests out of the large `lib.rs` facade.

Keep the generated WIT bindings and exported component in `lib.rs` while
preserving the existing Telegram runtime behaviour and test coverage.
