+++ channels-src/telegram/src/guest.rs
@@ -1,0 +1,298 @@
+use crate::exports::near::agent::channel::{
+    AgentResponse, ChannelConfig, Guest, HttpEndpointConfig, IncomingHttpRequest,
+    OutgoingHttpResponse, PollConfig, StatusType, StatusUpdate,
+};
+use crate::inbound::handle_update;
+use crate::polling::{fetch_updates, process_updates_response};
+use crate::send::{send_message, send_response};
+use crate::state::{
+    ALLOW_FROM_PATH, BOT_USERNAME_PATH, DM_POLICY_PATH, OWNER_ID_PATH, POLLING_STATE_PATH,
+    RESPOND_TO_ALL_GROUP_PATH,
+};
+use crate::status::{classify_status_update, TelegramStatusAction};
+use crate::types::{TelegramConfig, TelegramMessageMetadata};
+use crate::webhook::{delete_webhook, json_response, register_webhook};
+use crate::{channel_host, TelegramChannel};
+
+impl Guest for TelegramChannel {
+    fn on_start(config_json: String) -> Result<ChannelConfig, String> {
+        channel_host::log(
+            channel_host::LogLevel::Debug,
+            &format!("Telegram channel config: {}", config_json),
+        );
+
+        let config: TelegramConfig = serde_json::from_str(&config_json)
+            .map_err(|e| format!("Failed to parse config: {}", e))?;
+
+        channel_host::log(channel_host::LogLevel::Info, "Telegram channel starting");
+
+        if let Some(ref username) = config.bot_username {
+            channel_host::log(
+                channel_host::LogLevel::Info,
+                &format!("Bot username: @{}", username),
+            );
+        }
+
+        // Persist owner_id so subsequent callbacks (on_http_request, on_poll) can read it
+        if let Some(owner_id) = config.owner_id {
+            if let Err(e) = channel_host::workspace_write(OWNER_ID_PATH, &owner_id.to_string()) {
+                channel_host::log(
+                    channel_host::LogLevel::Error,
+                    &format!("Failed to persist owner_id: {}", e),
+                );
+            }
+            channel_host::log(
+                channel_host::LogLevel::Info,
+                &format!("Owner restriction enabled: user {}", owner_id),
+            );
+        } else {
+            // Clear any stale owner_id from a previous config
+            let _ = channel_host::workspace_write(OWNER_ID_PATH, "");
+            channel_host::log(
+                channel_host::LogLevel::Warn,
+                "No owner_id configured, bot is open to all users",
+            );
+        }
+
+        // Persist dm_policy and allow_from for DM pairing in handle_message
+        let dm_policy = config.dm_policy.as_deref().unwrap_or("pairing").to_string();
+        let _ = channel_host::workspace_write(DM_POLICY_PATH, &dm_policy);
+
+        let allow_from_json = serde_json::to_string(&config.allow_from.unwrap_or_default())
+            .unwrap_or_else(|_| "[]".to_string());
+        let _ = channel_host::workspace_write(ALLOW_FROM_PATH, &allow_from_json);
+
+        // Persist bot_username and respond_to_all_group_messages for group handling
+        let _ = channel_host::workspace_write(
+            BOT_USERNAME_PATH,
+            &config.bot_username.unwrap_or_default(),
+        );
+        let _ = channel_host::workspace_write(
+            RESPOND_TO_ALL_GROUP_PATH,
+            &config.respond_to_all_group_messages.to_string(),
+        );
+
+        // Mode: use polling if explicitly enabled, otherwise use webhooks when tunnel available.
+        let webhook_mode = config.tunnel_url.is_some() && !config.polling_enabled;
+
+        if webhook_mode {
+            channel_host::log(
+                channel_host::LogLevel::Info,
+                "Webhook mode enabled (tunnel configured)",
+            );
+
+            // Register webhook with Telegram API — propagate errors so a bad token
+            // causes activation to fail rather than silently succeeding.
+            if let Some(ref tunnel_url) = config.tunnel_url {
+                // Clear any stale webhook first to avoid 409 Conflict
+                let _ = delete_webhook();
+
+                channel_host::log(
+                    channel_host::LogLevel::Info,
+                    &format!("Registering webhook: {}/webhook/telegram", tunnel_url),
+                );
+
+                register_webhook(tunnel_url, config.webhook_secret.as_deref())
+                    .map_err(|e| format!("Failed to register webhook: {}", e))?;
+            }
+        } else {
+            channel_host::log(
+                channel_host::LogLevel::Info,
+                "Polling mode enabled (no tunnel configured)",
+            );
+
+            // Delete any existing webhook before polling. Telegram returns success
+            // when no webhook exists, so any error here (e.g. 401) means a bad token.
+            delete_webhook().map_err(|e| format!("Bot token validation failed: {}", e))?;
+        }
+
+        // Configure polling only if not in webhook mode
+        let poll = if !webhook_mode {
+            Some(PollConfig {
+                interval_ms: 30000, // 30 seconds minimum
+                enabled: true,
+            })
+        } else {
+            None
+        };
+
+        // Webhook secret validation is handled by the host
+        let require_secret = config.webhook_secret.is_some();
+
+        Ok(ChannelConfig {
+            display_name: "Telegram".to_string(),
+            http_endpoints: vec![HttpEndpointConfig {
+                path: "/webhook/telegram".to_string(),
+                methods: vec!["POST".to_string()],
+                require_secret,
+            }],
+            poll,
+        })
+    }
+
+    fn on_http_request(req: IncomingHttpRequest) -> OutgoingHttpResponse {
+        // Check if webhook secret validation passed (if required)
+        // The host validates X-Telegram-Bot-Api-Secret-Token header and sets secret_validated
+        // If require_secret was true in config but validation failed, secret_validated will be false
+        if !req.secret_validated {
+            // This means require_secret was set but the secret didn't match
+            // We still check the field even though the host should have already rejected invalid requests
+            // This is defense in depth
+            channel_host::log(
+                channel_host::LogLevel::Warn,
+                "Webhook request with invalid or missing secret token",
+            );
+            // Return 401 but Telegram will keep retrying, so this is just for logging
+            // In practice, the host should reject these before they reach us
+        }
+
+        // Parse the request body as UTF-8
+        let body_str = match std::str::from_utf8(&req.body) {
+            Ok(s) => s,
+            Err(_) => {
+                return json_response(400, serde_json::json!({"error": "Invalid UTF-8 body"}));
+            }
+        };
+
+        // Parse as Telegram Update
+        let update: crate::types::TelegramUpdate = match serde_json::from_str(body_str) {
+            Ok(u) => u,
+            Err(e) => {
+                channel_host::log(
+                    channel_host::LogLevel::Error,
+                    &format!("Failed to parse Telegram update: {}", e),
+                );
+                // Still return 200 to prevent Telegram from retrying
+                return json_response(200, serde_json::json!({"ok": true}));
+            }
+        };
+
+        // Handle the update
+        handle_update(update);
+
+        // Always respond 200 quickly (Telegram expects fast responses)
+        json_response(200, serde_json::json!({"ok": true}))
+    }
+
+    fn on_poll() {
+        let offset = match channel_host::workspace_read(POLLING_STATE_PATH) {
+            Some(s) => s.parse::<i64>().unwrap_or(0),
+            None => 0,
+        };
+
+        channel_host::log(
+            channel_host::LogLevel::Debug,
+            &format!("Polling getUpdates with offset {}", offset),
+        );
+
+        let headers_json = serde_json::json!({}).to_string();
+
+        match fetch_updates(offset, &headers_json) {
+            Ok(response) => process_updates_response(response, offset),
+            Err(e) => channel_host::log(
+                channel_host::LogLevel::Error,
+                &format!("getUpdates request failed: {}", e),
+            ),
+        }
+    }
+
+    fn on_respond(response: AgentResponse) -> Result<(), String> {
+        let metadata: TelegramMessageMetadata = serde_json::from_str(&response.metadata_json)
+            .map_err(|e| format!("Failed to parse metadata: {}", e))?;
+
+        send_response(metadata.chat_id, &response, Some(metadata.message_id))
+    }
+
+    fn on_broadcast(user_id: String, response: AgentResponse) -> Result<(), String> {
+        let chat_id: i64 = user_id
+            .parse()
+            .map_err(|e| format!("Invalid chat_id '{}': {}", user_id, e))?;
+
+        send_response(chat_id, &response, None)
+    }
+
+    fn on_status(update: StatusUpdate) {
+        let action = match classify_status_update(&update) {
+            Some(action) => action,
+            None => return,
+        };
+
+        // Parse chat_id from metadata
+        let metadata: TelegramMessageMetadata = match serde_json::from_str(&update.metadata_json) {
+            Ok(m) => m,
+            Err(_) => {
+                channel_host::log(
+                    channel_host::LogLevel::Debug,
+                    "on_status: no valid Telegram metadata, skipping status update",
+                );
+                return;
+            }
+        };
+
+        match action {
+            TelegramStatusAction::Typing => {
+                // POST /sendChatAction with action "typing"
+                let payload = serde_json::json!({
+                    "chat_id": metadata.chat_id,
+                    "action": "typing"
+                });
+
+                let payload_bytes = match serde_json::to_vec(&payload) {
+                    Ok(b) => b,
+                    Err(_) => return,
+                };
+
+                let headers = serde_json::json!({
+                    "Content-Type": "application/json"
+                });
+
+                let result = channel_host::http_request(&channel_host::HttpRequestParams {
+                    method: "POST".to_string(),
+                    url: "https://api.telegram.org/bot{TELEGRAM_BOT_TOKEN}/sendChatAction"
+                        .to_string(),
+                    headers_json: headers.to_string(),
+                    body: Some(payload_bytes.clone()),
+                    timeout_ms: None,
+                });
+
+                if let Err(e) = result {
+                    channel_host::log(
+                        channel_host::LogLevel::Debug,
+                        &format!("sendChatAction failed: {}", e),
+                    );
+                }
+            }
+            TelegramStatusAction::Notify(prompt) => {
+                // Send user-visible status updates for actionable events.
+                if let Err(first_err) =
+                    send_message(metadata.chat_id, &prompt, Some(metadata.message_id), None)
+                {
+                    channel_host::log(
+                        channel_host::LogLevel::Warn,
+                        &format!(
+                            "Failed to send status reply ({}), retrying without reply context",
+                            first_err
+                        ),
+                    );
+
+                    if let Err(retry_err) = send_message(metadata.chat_id, &prompt, None, None) {
+                        channel_host::log(
+                            channel_host::LogLevel::Debug,
+                            &format!(
+                                "Failed to send status message without reply context: {}",
+                                retry_err
+                            ),
+                        );
+                    }
+                }
+            }
+        }
+    }
+
+    fn on_shutdown() {
+        channel_host::log(
+            channel_host::LogLevel::Info,
+            "Telegram channel shutting down",
+        );
+    }
+}
