+++ channels-src/telegram/src/state.rs
@@ -1,0 +1,13 @@
+pub(crate) const POLLING_STATE_PATH: &str = "state/last_update_id";
+
+pub(crate) const OWNER_ID_PATH: &str = "state/owner_id";
+
+pub(crate) const DM_POLICY_PATH: &str = "state/dm_policy";
+
+pub(crate) const ALLOW_FROM_PATH: &str = "state/allow_from";
+
+pub(crate) const CHANNEL_NAME: &str = "telegram";
+
+pub(crate) const BOT_USERNAME_PATH: &str = "state/bot_username";
+
+pub(crate) const RESPOND_TO_ALL_GROUP_PATH: &str = "state/respond_to_all_group_messages";
